--- a/actioncable/Rakefile
+++ b/actioncable/Rakefile
@@ -65,7 +65,7 @@ namespace :assets do
     end
 
     print "[verify] #{dir} can be required as a module "
-    stdout, stderr, status = Open3.capture3("node", "--print", "window = {}; require('#{dir}');")
+    _, stderr, status = Open3.capture3("node", "--print", "window = {}; require('#{dir}');")
     if status.success?
       puts "[OK]"
     else