--- a/actionpack/lib/abstract_controller/base.rb
+++ b/actionpack/lib/abstract_controller/base.rb
@@ -14,8 +14,16 @@ module AbstractController
   # expected to provide their own +render+ method, since rendering means
   # different things depending on the context.
   class Base
+    ##
+    # Returns the HTTP response sent by the controller
     attr_internal :response_body
+
+    ##
+    # Returns the name of the action this controller is processing
     attr_internal :action_name
+
+    ##
+    # Returns the formats processed by the controller
     attr_internal :formats
 
     include ActiveSupport::Configurable