--- a/actionpack/lib/action_dispatch/system_testing/test_helpers/setup_and_teardown.rb
+++ b/actionpack/lib/action_dispatch/system_testing/test_helpers/setup_and_teardown.rb
@@ -2,7 +2,12 @@ module ActionDispatch
   module SystemTesting
     module TestHelpers
       module SetupAndTeardown # :nodoc:
-        DEFAULT_HOST = "127.0.0.1"
+        DEFAULT_HOST = "http://127.0.0.1"
+
+        def host!(host)
+          super
+          Capybara.app_host = host
+        end
 
         def before_setup
           host! DEFAULT_HOST