--- a/actionpack/test/dispatch/system_testing/system_test_case_test.rb
+++ b/actionpack/test/dispatch/system_testing/system_test_case_test.rb
@@ -19,3 +19,15 @@ class SetDriverToSeleniumTest < DrivenBySeleniumWithChrome
     assert_equal :selenium, Capybara.current_driver
   end
 end
+
+class SetHostTest < DrivenByRackTest
+  test "sets default host" do
+    assert_equal "http://127.0.0.1", Capybara.app_host
+  end
+
+  test "overrides host" do
+    host! "http://example.com"
+
+    assert_equal "http://example.com", Capybara.app_host
+  end
+end