… .travis.yml
… Gemfile.lock
… actionpack/CHANGELOG.md
… actionview/CHANGELOG.md
… actionview/app/assets/javascripts/README.md
… actionview/test/template/form_options_helper_test.rb
… actionview/test/template/url_helper_test.rb
… activerecord/CHANGELOG.md
… activerecord/test/cases/associations/callbacks_test.rb
… activerecord/test/cases/associations/has_many_through_associations_test.rb
… activerecord/test/cases/associations/inverse_associations_test.rb
… activerecord/test/cases/cache_key_test.rb
… activerecord/test/cases/integration_test.rb
… activerecord/test/cases/migration/references_statements_test.rb
… activerecord/test/cases/nested_attributes_test.rb
… activerecord/test/cases/relation/mutation_test.rb
… activesupport/CHANGELOG.md
… activesupport/test/core_ext/duration_test.rb
… guides/source/action_controller_overview.md
… guides/source/action_mailer_basics.md
… guides/source/action_view_overview.md
… guides/source/active_job_basics.md
… guides/source/api_app.md
… guides/source/api_documentation_guidelines.md
… guides/source/association_basics.md
… guides/source/getting_started.md
… guides/source/layouts_and_rendering.md
… guides/source/nested_model_forms.md
… guides/source/profiling.md
… guides/source/rails_on_rack.md