@jamiemccarthy 2017-05-19
This looks :100: Lovitt. Great work and great writeup. Let's submit it!