@thenonsequitur 2017-05-19
:+1: This is awesome! Looks ready for submission to me too.