@lovitt 2017-05-19
OK @jamiemccarthy and @thenonsequitur, I realized today that we never actually applied the `#deep_dup` patch to Chorus. After discussing with Jamie, current plan is to:

1) Deploy that patch today: https://github.com/voxmedia/sbn/pull/5610
2) Let it run for a few days, ensure there are no issues
3) Early next week, do a canary deploy of a branch that reverts **all three** of these optimizations
4) Gather data for a few hours, and then measure performance of unpatched compred to patched, and use those metrics in this PR when describing our real-world speedup. 

I considered just removing that real-world results paragraph from this PR so that we could ship it and get on with our lives, but Jamie convinced me that taking a bit more time here will be worth it -- we'll be able to provide a more accurate indication of the impact of this patch to the Rails community (and to ourselves).