@lovitt 2017-05-30
Merged into rails/rails: https://github.com/rails/rails/pull/29215